Extract TUF Updater into tuf/
Separating the general TUF pieces from the sigstore-specific bits
